refactor: Trade 도메인 분리
Trade 도메인 분리
