--- src/main/java/kr/co/fastcampus/yanabada/domain/payment/entity/enums/TradeRole.java
@@ -1,6 +1,0 @@
-package kr.co.fastcampus.yanabada.domain.payment.entity.enums;
-
-public enum TradeRole {
-    SELLER,
-    BUYER
-}
