+++ src/main/java/kr/co/fastcampus/yanabada/domain/product/service/ProductService.java
@@ -33,12 +33,9 @@ import kr.co.fastcampus.yanabada.domain.order.entity.enums.OrderStatus;
 import kr.co.fastcampus.yanabada.domain.order.entity.enums.PaymentType;
 import kr.co.fastcampus.yanabada.domain.order.repository.OrderRepository;
 import kr.co.fastcampus.yanabada.domain.payment.entity.AdminPayment;
-import kr.co.fastcampus.yanabada.domain.payment.entity.Trade;
 import kr.co.fastcampus.yanabada.domain.payment.entity.YanoljaPay;
 import kr.co.fastcampus.yanabada.domain.payment.entity.YanoljaPayHistory;
-import kr.co.fastcampus.yanabada.domain.payment.entity.enums.TradeStatus;
 import kr.co.fastcampus.yanabada.domain.payment.repository.AdminPaymentRepository;
-import kr.co.fastcampus.yanabada.domain.payment.repository.TradeRepository;
 import kr.co.fastcampus.yanabada.domain.payment.repository.YanoljaPayHistoryRepository;
 import kr.co.fastcampus.yanabada.domain.payment.repository.YanoljaPayRepository;
 import kr.co.fastcampus.yanabada.domain.product.dto.request.ProductPatchRequest;
@@ -52,6 +49,9 @@ import kr.co.fastcampus.yanabada.domain.product.dto.response.ProductSummaryPageR
 import kr.co.fastcampus.yanabada.domain.product.entity.Product;
 import kr.co.fastcampus.yanabada.domain.product.entity.enums.ProductStatus;
 import kr.co.fastcampus.yanabada.domain.product.repository.ProductRepository;
+import kr.co.fastcampus.yanabada.domain.trade.entity.Trade;
+import kr.co.fastcampus.yanabada.domain.trade.entity.enums.TradeStatus;
+import kr.co.fastcampus.yanabada.domain.trade.repository.TradeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
