+++ src/main/java/kr/co/fastcampus/yanabada/domain/trade/dto/response/PurchaseTradeInfoResponse.java
@@ -1,9 +1,9 @@
-package kr.co.fastcampus.yanabada.domain.payment.dto.response;
+package kr.co.fastcampus.yanabada.domain.trade.dto.response;
 
 import java.time.LocalDateTime;
 import kr.co.fastcampus.yanabada.domain.order.entity.enums.PaymentType;
-import kr.co.fastcampus.yanabada.domain.payment.entity.Trade;
-import kr.co.fastcampus.yanabada.domain.payment.entity.enums.TradeStatus;
+import kr.co.fastcampus.yanabada.domain.trade.entity.Trade;
+import kr.co.fastcampus.yanabada.domain.trade.entity.enums.TradeStatus;
 import kr.co.fastcampus.yanabada.domain.product.dto.response.ProductSummaryResponse;
 import kr.co.fastcampus.yanabada.domain.product.entity.Product;
 import lombok.Builder;
