+++ src/main/java/kr/co/fastcampus/yanabada/domain/trade/dto/response/TradeIdResponse.java
@@ -1,6 +1,6 @@
-package kr.co.fastcampus.yanabada.domain.payment.dto.response;
+package kr.co.fastcampus.yanabada.domain.trade.dto.response;
 
-import kr.co.fastcampus.yanabada.domain.payment.entity.Trade;
+import kr.co.fastcampus.yanabada.domain.trade.entity.Trade;
 import lombok.Builder;
 
 @Builder
