+++ src/main/java/kr/co/fastcampus/yanabada/domain/trade/entity/enums/TradeRole.java
@@ -1,0 +1,6 @@
+package kr.co.fastcampus.yanabada.domain.trade.entity.enums;
+
+public enum TradeRole {
+    SELLER,
+    BUYER
+}
