+++ src/main/java/kr/co/fastcampus/yanabada/domain/trade/entity/enums/TradeStatus.java
@@ -1,4 +1,4 @@
-package kr.co.fastcampus.yanabada.domain.payment.entity.enums;
+package kr.co.fastcampus.yanabada.domain.trade.entity.enums;
 
 public enum TradeStatus {
     WAITING, COMPLETED, CANCELED, REJECTED
