+++ src/main/java/kr/co/fastcampus/yanabada/domain/trade/repository/TradeRepository.java
@@ -1,12 +1,12 @@
-package kr.co.fastcampus.yanabada.domain.payment.repository;
+package kr.co.fastcampus.yanabada.domain.trade.repository;
 
 import java.util.List;
 import java.util.Optional;
 import kr.co.fastcampus.yanabada.common.exception.TradeNotFoundException;
 import kr.co.fastcampus.yanabada.domain.member.entity.Member;
-import kr.co.fastcampus.yanabada.domain.payment.entity.Trade;
-import kr.co.fastcampus.yanabada.domain.payment.entity.enums.TradeStatus;
 import kr.co.fastcampus.yanabada.domain.product.entity.Product;
+import kr.co.fastcampus.yanabada.domain.trade.entity.Trade;
+import kr.co.fastcampus.yanabada.domain.trade.entity.enums.TradeStatus;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
