+++ src/main/java/kr/co/fastcampus/yanabada/domain/trade/service/TradeService.java
@@ -1,14 +1,14 @@
-package kr.co.fastcampus.yanabada.domain.payment.service;
+package kr.co.fastcampus.yanabada.domain.trade.service;
 
 import static kr.co.fastcampus.yanabada.domain.order.entity.enums.PaymentType.YANOLJA_PAY;
 import static kr.co.fastcampus.yanabada.domain.payment.entity.enums.ContentsType.PURCHASE;
 import static kr.co.fastcampus.yanabada.domain.payment.entity.enums.ContentsType.REFUND;
 import static kr.co.fastcampus.yanabada.domain.payment.entity.enums.ContentsType.SALE;
-import static kr.co.fastcampus.yanabada.domain.payment.entity.enums.TradeRole.BUYER;
-import static kr.co.fastcampus.yanabada.domain.payment.entity.enums.TradeRole.SELLER;
-import static kr.co.fastcampus.yanabada.domain.payment.entity.enums.TradeStatus.WAITING;
 import static kr.co.fastcampus.yanabada.domain.payment.entity.enums.TransactionType.DEPOSIT;
 import static kr.co.fastcampus.yanabada.domain.payment.entity.enums.TransactionType.WITHDRAW;
+import static kr.co.fastcampus.yanabada.domain.trade.entity.enums.TradeRole.BUYER;
+import static kr.co.fastcampus.yanabada.domain.trade.entity.enums.TradeRole.SELLER;
+import static kr.co.fastcampus.yanabada.domain.trade.entity.enums.TradeStatus.WAITING;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
@@ -30,27 +30,27 @@ import kr.co.fastcampus.yanabada.domain.order.entity.Order;
 import kr.co.fastcampus.yanabada.domain.order.entity.enums.OrderStatus;
 import kr.co.fastcampus.yanabada.domain.order.entity.enums.PaymentType;
 import kr.co.fastcampus.yanabada.domain.order.repository.OrderRepository;
-import kr.co.fastcampus.yanabada.domain.payment.dto.request.TradeSaveRequest;
-import kr.co.fastcampus.yanabada.domain.payment.dto.response.ApprovalTradeInfoResponse;
-import kr.co.fastcampus.yanabada.domain.payment.dto.response.ApprovalTradePageResponse;
-import kr.co.fastcampus.yanabada.domain.payment.dto.response.ApprovalTradeSummaryResponse;
-import kr.co.fastcampus.yanabada.domain.payment.dto.response.PurchaseTradeInfoResponse;
-import kr.co.fastcampus.yanabada.domain.payment.dto.response.PurchaseTradePageResponse;
-import kr.co.fastcampus.yanabada.domain.payment.dto.response.PurchaseTradeSummaryResponse;
-import kr.co.fastcampus.yanabada.domain.payment.dto.response.TradeIdResponse;
 import kr.co.fastcampus.yanabada.domain.payment.entity.AdminPayment;
-import kr.co.fastcampus.yanabada.domain.payment.entity.Trade;
 import kr.co.fastcampus.yanabada.domain.payment.entity.YanoljaPay;
 import kr.co.fastcampus.yanabada.domain.payment.entity.YanoljaPayHistory;
-import kr.co.fastcampus.yanabada.domain.payment.entity.enums.TradeRole;
-import kr.co.fastcampus.yanabada.domain.payment.entity.enums.TradeStatus;
 import kr.co.fastcampus.yanabada.domain.payment.repository.AdminPaymentRepository;
-import kr.co.fastcampus.yanabada.domain.payment.repository.TradeRepository;
 import kr.co.fastcampus.yanabada.domain.payment.repository.YanoljaPayHistoryRepository;
 import kr.co.fastcampus.yanabada.domain.payment.repository.YanoljaPayRepository;
 import kr.co.fastcampus.yanabada.domain.product.entity.Product;
 import kr.co.fastcampus.yanabada.domain.product.entity.enums.ProductStatus;
 import kr.co.fastcampus.yanabada.domain.product.repository.ProductRepository;
+import kr.co.fastcampus.yanabada.domain.trade.dto.request.TradeSaveRequest;
+import kr.co.fastcampus.yanabada.domain.trade.dto.response.ApprovalTradeInfoResponse;
+import kr.co.fastcampus.yanabada.domain.trade.dto.response.ApprovalTradePageResponse;
+import kr.co.fastcampus.yanabada.domain.trade.dto.response.ApprovalTradeSummaryResponse;
+import kr.co.fastcampus.yanabada.domain.trade.dto.response.PurchaseTradeInfoResponse;
+import kr.co.fastcampus.yanabada.domain.trade.dto.response.PurchaseTradePageResponse;
+import kr.co.fastcampus.yanabada.domain.trade.dto.response.PurchaseTradeSummaryResponse;
+import kr.co.fastcampus.yanabada.domain.trade.dto.response.TradeIdResponse;
+import kr.co.fastcampus.yanabada.domain.trade.entity.Trade;
+import kr.co.fastcampus.yanabada.domain.trade.entity.enums.TradeRole;
+import kr.co.fastcampus.yanabada.domain.trade.entity.enums.TradeStatus;
+import kr.co.fastcampus.yanabada.domain.trade.repository.TradeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
